+++ resources/views/layouts/boarding.blade.php
@@ -1,0 +1,9 @@
+@extends('layouts.base')
+@section('body')
+<main class="min-h-screen hero">
+    <div class="hero-content">
+        {{ $slot }}
+    </div>
+</main>
+@parent
+@endsection
