+++ resources/views/layouts/simple.blade.php
@@ -1,0 +1,7 @@
+@extends('layouts.base')
+@section('body')
+    @parent
+    <main>
+        {{ $slot }}
+    </main>
+@endsection
